Detect non-Mac and non-Windows as Unix to support BSDs next to Linux
Building on OpenBSD fails since `showCertificate()` attempts to use
the Windows-specific `addTempFile()` function.

This is because the current code assumes the target platform to be one
of either MacOS, Windows or Linux.

`Q_OS_MAC` and `Q_OS_WIN` requiring a number of platform specific code
aside, the current code under `Q_OS_LINUX` largely works unmodified on
other `Q_OS_UNIX` platforms such as OpenBSD.

More fixes are required, e.g. to PKCS11 module loading, but this is the
only one required to build DigiDoc4 Client as-is on OpenBSD.

Signed-off-by: Klemens Nanni <klemens@posteo.de>