Restrict Linux-specific modules to Linux
`opensc-pkcs11.so` is practically available on every platform and the
relative filename causes it to be searched for in runtime library path
controlled by the platform, making it work by default on the BSDs for
example.

Other modules however use absolute paths and/or specific keys, i.e. the
code assumes specific platforms and/or modules origins.

On officially unsupported platforms, e.g. OpenBSD, these specific
modules will never be found, so only build them on Linux.

Signed-off-by: Klemens Nanni <klemens@posteo.de>